more fixing for correct usage of TMP OUT FOLDER

+++ docker-test.sh
@@ -31,19 +31,19 @@ fi
 #run the built container
 docker run --rm \
   --name emo-bon_${dckr_cnt_nm} \
-  --volume ${TMPDIR}:/rocrateroot \
+  --volume ${TEST_OUTPUTFOLDER}:/rocrateroot \
   --env ARUP_WORK=test-work.yml \
   --env ARUP_TEMPLATES=test-templates \
   --env SOURCE_MAT_ID='test_sm_id' \
   ${dckr_img_nm}
 
 #verify the output
-test -f ${TMPDIR}/test-output.ttl || (echo "mising output file" && exit 1)
-test -f ${TMPDIR}/test-output2.ttl || (echo "missing 2nd output file" && exit 1)
-diff ${TMPDIR}/test-output.ttl ${TMPDIR}/test-output2.ttl || (echo "unexpected diff between output files" && exit 1)
+test -f ${TEST_OUTPUTFOLDER}/test-output.ttl || (echo "mising output file" && exit 1)
+test -f ${TEST_OUTPUTFOLDER}/test-output2.ttl || (echo "missing 2nd output file" && exit 1)
+diff ${TEST_OUTPUTFOLDER}/test-output.ttl ${TEST_OUTPUTFOLDER}/test-output2.ttl || (echo "unexpected diff between output files" && exit 1)
 ttl=$(which ttl)  # look for ttl validator
 if [[ -x "${ttl}" ]]; then
-  ${ttl} ${TMPDIR}/test-output.ttl || (echo "ttl validation failed" && exit 1)
+  ${ttl} ${TEST_OUTPUTFOLDER}/test-output.ttl || (echo "ttl validation failed" && exit 1)
   # no need to check the 2nd file, it is the same as the first
 else
   echo "**WARN** ttl validator not found, skipping validation guarantee."
@@ -51,5 +51,10 @@ fi
 
 #say bye and clean up
 echo "test passed, cleaning up"
-rm -rf ${TMPDIR}
+if [[ -z ${TMPDIR} ]]; then
+  echo "no tempfolder to remove"
+else
+  echo "cleanup temp folder"
+  rm -rf ${TMPDIR}
+fi
 exit 0
